added add new movie functionality
