+++ src/components/newMovie.jsx
@@ -2,10 +2,10 @@ import React from "react";
 import axios from "axios";
 import { DATA_ENDPOINT } from "../config";
 
-function NewMovie() {
+function NewMovie({ setMovies }) {
   const handleSubmit = (e) => {
     e.preventDefault();
-    console.log(e.target.form.title.value);
+    console.log(`adding: ${e.target.form.title.value}`);
 
     axios
       .post(`${DATA_ENDPOINT}/movies`, {
@@ -16,13 +16,21 @@ function NewMovie() {
       })
       .then((response) => {
         console.log(response);
+        e.target.form.title.value = "";
+        e.target.form.stock.value = "";
+        e.target.form.price.value = "";
+
+        axios.get(`${DATA_ENDPOINT}/movies`).then((response) => {
+          console.log("fetch after adding");
+          setMovies(response.data);
+        });
       });
   };
 
   return (
     <form>
       <h4>Add new movie</h4>
-      <label htmlFor="title">Add new movie: </label>
+      <label htmlFor="title">Add new movie:</label>
       <input
         type="text"
         className="form-control"
@@ -30,7 +38,9 @@ function NewMovie() {
         placeholder="movie title"
       />
 
-      <label htmlFor="stock">number in stock: </label>
+      <label className="mt-2" htmlFor="stock">
+        number in stock:
+      </label>
       <input
         type="text"
         className="form-control"
@@ -38,7 +48,9 @@ function NewMovie() {
         placeholder="items in stock"
       />
 
-      <label htmlFor="price">daily rental: </label>
+      <label className="mt-2" htmlFor="price">
+        daily rental:
+      </label>
       <input
         type="text"
         className="form-control"
@@ -48,7 +60,7 @@ function NewMovie() {
 
       <button
         type="submit"
-        className="btn btn-primary mt-2"
+        className="btn btn-primary mt-3"
         onClick={(e) => handleSubmit(e)}
       >
         Submit
